refactor(encrypt): move encryption logic into their own file
