+++ lib/encrypt.js
@@ -2,15 +2,13 @@
 
 module.exports = encryptDoc
 
-var aes = require('browserify-aes')
 var includes = require('lodash/includes')
 var randomBytes = require('randombytes')
 var Promise = require('lie')
 var uuid = require('uuid').v4
 
+var encrypt = require('./helpers/encrypt-core')
 var ignore = require('./utils/ignore.js')
-
-var subtle = global.crypto && global.crypto.subtle
 
 // Mostly copied from https://github.com/calvinmetcalf/crypto-pouch/blob/master/index.js#L130
 function encryptDoc (state, doc, prefix) {
@@ -52,82 +50,11 @@ function encryptDoc (state, doc, prefix) {
   }
 
   var data = Buffer.from(JSON.stringify(encryptDoc))
-  return checkBrowser()
+  return encrypt(state.key, nonce, data, Buffer.from(outDoc._id))
 
-    // Original implementation from:
-    // https://github.com/calvinmetcalf/native-crypto/blob/master/browser/encrypt.js
-    .then(function (supportsBrowserCrypto) {
-      // for when the browser supports crypto.subtle
-      if (supportsBrowserCrypto) {
-        return subtle.importKey('raw', state.key, { name: 'AES-GCM' }, true, ['encrypt'])
-
-          .then(function (key) {
-            var options = {
-              name: 'AES-GCM',
-              iv: nonce,
-              additionalData: Buffer.from(outDoc._id)
-            }
-
-            return subtle.encrypt(options, key, data)
-          })
-
-          .then(function (response) {
-            var buffy = Buffer.from(response)
-            outDoc.tag = buffy.slice(-16).toString('hex')
-            outDoc.data = buffy.slice(0, -16).toString('hex')
-            return outDoc
-          })
-      } else {
-        // fall back to browserify-aes
-        var cipher = aes.createCipheriv('aes-256-gcm', state.key, nonce)
-        cipher.setAAD(Buffer.from(outDoc._id))
-
-        var output = cipher.update(data)
-        cipher.final()
-        var tag = cipher.getAuthTag()
-
-        outDoc.tag = tag.toString('hex')
-        outDoc.data = output.toString('hex')
-        return outDoc
-      }
-    })
-}
-
-var canUseBrowserCrypto = null
-
-function checkBrowser () {
-  if (global.process && !global.process.browser) {
-    return Promise.resolve(false)
-  }
-  if (!subtle || !subtle.importKey || !subtle.encrypt) {
-    return Promise.resolve(false)
-  }
-  if (canUseBrowserCrypto != null) {
-    return Promise.resolve(canUseBrowserCrypto)
-  }
-
-  var zeroBuffy = Buffer.alloc(32, 0)
-  var ivFaith = Buffer.alloc(12, 0)
-
-  return subtle.importKey('raw', zeroBuffy.buffer, { name: 'AES-GCM' }, true, ['encrypt'])
-
-    .then(function (key) {
-      return subtle.encrypt(
-        { name: 'AES-GCM', iv: ivFaith },
-        key,
-        zeroBuffy.buffer
-      )
-    })
-
-    .then(function (res) {
-      canUseBrowserCrypto = Buffer.from(res)
-        .toString('base64') === 'zqdAPU1ga24HTsXTuvOdGHJgA8o3pip00aL1jnUGNY7R0whMmaqKn9q7PoPrKMFd'
-
-      return canUseBrowserCrypto
-    })
-
-    .catch(function () {
-      canUseBrowserCrypto = false
-      return canUseBrowserCrypto
+    .then(function (result) {
+      outDoc.tag = result.tag
+      outDoc.data = result.data
+      return outDoc
     })
 }
