+++ lib/setup.js
@@ -10,7 +10,7 @@ var createResetKeys = require('./helpers/create-reset-keys')
 module.exports = setup
 
 /**
- * Setup function for first run and generating a key and salt. It dosn't unlock!
+ * Setup function for first run and generating a key and salt. It doesn't unlock!
  *
  * @param  {Object}    store        instance of a hoodie client store
  * @param  {Object}    state        crypto config
@@ -86,7 +86,7 @@ function setup (store, state, password, salt) {
       var obj = objs[0]
       var resetKeys = objs[1]
 
-      // if there was an deleted salt doc: undelete it
+      // if there was an deleted salt doc: un-delete it
       if (obj._deleted) {
         obj._deleted = false
         return store.updateOrAdd(obj)
